Removed Dependency of sqlite3 unixepoch to avoid conflicts with older versions

+++ src/salduba/ib_tws_proxy/backing_db/schema/meta.sql
@@ -1,6 +1,6 @@
 create table if not exists db_info (
   version integer primary key,
-  version_time integer,
+  version_time bigint,
   at integer not NULL,
   current BOOLEAN not NULL
 );
